set up initial should render tests
