+++ src/app/__tests__/Dashboard/Folder.test.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { shallow } from 'enzyme';
+import Folder from '../../Dashboard/Folder';
+
+describe('<Folder />', () => {
+  let wrapper;
+  beforeEach(() => {
+    wrapper = shallow(<Folder />);
+  });
+
+  it('renders without crashing', () => {});
+});
